+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 Version 2.1.0-dev
 - allowed customization of test suit creation in automated tests runner
 - made Job::$name, Job::$params and Job::$shouldFail readable
-- deprecated Environment::NAME, Environment::printLine(), Environment::getShouldFail(), Environment::setShouldFail(), Environment::$shouldFail, Environment::VERSION, Environment::$currentJob, Environment::printInfo(), Environment::getSkipped(), Environment::printResults()
+- deprecated Environment
+- possible BC break: only public methods in TestCase whose name starts with *test* are now considered tests
 
 Version 2.0.1
 - changed default value for attributes to true